Module varu64::nb

Nonblocking encoding and decoding.

Structs

Decoder
State for the nonblocking decoding.

Encoder
State for the nonblocking encoding.

LengthValueDecoder
State for decoding a VarU64 followed by that many bytes into a Vec<u8>.

LengthValueEncoder
State for encoding some bytes, preceded by a VarU64 indicating their length

LengthValueLimitDecoder
State for decoding a VarU64 followed by that many bytes into a Vec<u8>, erroring if the VarU64 is greater than a limit.

Enums

DecodeError
Everything that can go wrong when decoding data.

DecodeLimitError
Everything that can go wrong when decoding a length-value with a limit.