Crate varu64

Implementation of the varu64 format in rust.

Modules

nb

Nonblocking encoding and decoding.

Enums

DecodeError

Everything that can go wrong when decoding a varu64.

Functions

decode

Decode a u64 from the input buffer, returning the number and the remaining bytes.

encode

Encodes n into the output buffer, returning how many bytes have been written.

encode_write

Encodes n into the writer, returning how many bytes have been written.

encoding_length

Return how many bytes the encoding of n will take up.